+++ packages/data-access/src/utils/crypto.ts
@@ -1,29 +1,12 @@
 import crypto from 'crypto';
 import { promisify } from 'util';
-import { cloneDeep, isPlainObject, uniq } from '@terascope/utils';
 
 const randomBytesAsync = promisify(crypto.randomBytes);
 const pbkdf2Async = promisify(crypto.pbkdf2);
 
-export function addDefaults(source: object, from: object = {}) {
-    const output = cloneDeep(source);
-    const _mapping = cloneDeep(from);
-
-    for (const [key, val] of Object.entries(_mapping)) {
-        if (output[key] != null) {
-            if (isPlainObject(val)) {
-                output[key] = Object.assign(output[key], val);
-            } else if (Array.isArray(val)) {
-                output[key] = uniq(output[key].concat(val));
-            } else {
-                output[key] = val;
-            }
-        }
-    }
-
-    return output;
-}
-
+/**
+ * Generate a API Token
+*/
 export async function generateAPIToken(hash: string, username: string) {
     const shasum = crypto.createHash('sha1');
     const buf = await randomBytesAsync(128);
@@ -33,11 +16,17 @@ export async function generateAPIToken(hash: string, username: string) {
     return shasum.update(str).digest('hex');
 }
 
+/**
+ * Generate a random salt
+*/
 export async function generateSalt() {
     const buf = await randomBytesAsync(32);
     return buf.toString('hex');
 }
 
+/**
+ * Generate a secure password hash
+*/
 export async function generatePasswordHash(password: string, salt: string) {
     const buf = await pbkdf2Async(password, salt, 25000, 512, 'sha1');
     return buf.toString('hex');
