+++ packages/data-access/src/utils/index.ts
@@ -1,9 +1,2 @@
-import nanoid from 'nanoid/async';
-
-export function makeISODate() {
-    return new Date().toISOString();
-}
-
-export function makeId() {
-    return nanoid(12);
-}
+export * from './crypto';
+export * from './misc';
